requirements.txt
asgiref==3.11.1
cryptography==46.0.2
Django==6.0.3
django-allauth==65.15.0
django-unfold==0.84.0
Faker==40.8.0
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.5
tzdata==2025.3
